\name{bootstrap}
\alias{bootstrap}
\alias{extractBOOT}
\title{
Do bootstrap and extract bootstrap results
}
\description{
Do bootstrap and extract bootstrap results.
}
\usage{
bootstrap(object, ...)
extractBOOT(object, ...)
}
\arguments{
  \item{object}{
a fitted model object
}
  \item{\dots}{
other arguments
}
}
\value{
\code{bootstrap} performs bootstrap.

\code{extractBOOT} is used to extract bootstrap results.
}
\author{
Peter Solymos
}
\keyword{ htest }
