\name{josm}
\alias{josm}
\docType{data}
\title{
Bird point count data from Alberta's Oil Sands region
}
\description{
Data set used in Mahon et al. 2016 and Mahon et al. 2019.
}
\usage{data("josm")}
\format{
  A list with elements \code{surveys}, \code{species}, and \code{counts}.
}
\source{
Mahon et al. 2016 and Mahon et al. 2019.
}
\references{
Mahon, C. L., Holloway, G., Solymos, P., Cumming, S. G., Bayne, E. M., Schmiegelow, F. K. A., Song, S. J., 2016.
Community structure and niche characteristics of upland and lowland western boreal birds at multiple spatial scales
\emph{Forest Ecology and Management} \bold{361}:99-116.
<doi:10.1016/j.foreco.2015.11.007>

Mahon, C. L., Holloway, G., Bayne, E. M., Toms, J. D., 2019.
Additive and interactive cumulative effects on boreal landbirds: winners and losers in a multi-stressor landscape
\emph{Ecological Applications} \bold{29}:e01895.
<doi:10.1002/eap.1895>
}
\examples{
data(josm)
str(josm)
}
\keyword{datasets}
