\name{lhreg}
\alias{lhreg}
\docType{data}
\title{
Trait data and phylogenetic correlation for 141 bird species
}
\description{
Data set used in Solymos et al. 2018.
}
\usage{data("lhreg")}
\format{
  A list with elements \code{traits}, and \code{phylo_cor}.
}
\source{
Solymos et al. 2018.
}
\references{
Solymos, P., Matsuoka, S. M., Stralberg, D., Barker, N. K. S., and Bayne, E. M., 2018. 
Phylogeny and species traits predict bird detectability. 
\emph{Ecography} \bold{41}:1595-1603.
<doi:10.1111/ecog.03415>
}
\examples{
data(lhreg)
str(lhreg)
}
\keyword{datasets}
