\name{load_BAM_QPAD}
\alias{load_BAM_QPAD}
\title{
Load BAM QPAD parameter estimates and support functions
}
\description{
Load BAM QPAD parameter estimates and support functions.
}
\usage{
load_BAM_QPAD(version)
}
\arguments{
  \item{version}{
version of the BAM QPAD estimates. List of selection is
provided if missing.
}
}
\details{
The \code{load_BAM_QPAD} function in the 'detect' package is deprecated.
Use the \code{load_BAM_QPAD} function 'QPAD' package instead.
See \url{https://github.com/psolymos/QPAD} for more information.
}
\value{
It returns a message.
}
\references{
Solymos, P., Matsuoka, S. M., Bayne, E. M., Lele, S. R., Fontaine, P.,
Cumming, S. G., Stralberg, D., Schmiegelow, F. K. A. & Song, S. J., 2013.
Calibrating indices of avian density from non-standardized survey data:
making the most of a messy situation.
\emph{Methods in Ecology and Evolution}, \bold{4}, 1047--1058.
<doi:10.1111/2041-210X.12106>

Supporting info, including a tutorial for the above paper:
\url{https://github.com/psolymos/QPAD/tree/master/inst/doc/v2}
}
\author{
Peter Solymos
}
\keyword{ misc }
