\name{detect-package}
\alias{detect-package}
\alias{detect}
\docType{package}
\title{
Analyzing Single Visit Data with Detection Error
}
\description{
The package implements zero inflated Binomial model
  for single visit site occupancy and Binomial-ZIP mixture model 
  for single visit count data when the probability of detection
  is less than one. Provides hands-on formula interface with inferences
  based on numerical optimization and data cloning.
  Development was funded by the Alberta Biodiversity Monitoring Institute.
}
\details{
\code{\link{svocc}}: single visit occupancy model.

\code{\link{svabu}}: single visit abundance model.
}
\author{
Peter Solymos, Monica Moreno

Maintainer: Peter Solymos <solymos@ualberta.ca>
}
\references{
Lele et al.

Solymos et al.
}
\keyword{ package }
\examples{
## how to open tutorial/vignette ???
}
