\name{detect-package}
\alias{detect-package}
\alias{detect}
\docType{package}
\title{
Analyzing Wildlife Data with Detection Error
}
\description{
  The package implements models to analyze
  site occupancy and count data models with detection error.
  Development was funded by the Alberta Biodiversity Monitoring Institute
  \url{http://www.abmi.ca/}.
}
\details{
\code{\link{svocc}}: single visit occupancy model.

\code{\link{svabu}}: single visit abundance model.
}
\author{
Peter Solymos, Monica Moreno, Subhash R Lele

Maintainer: Peter Solymos <solymos@ualberta.ca>
}
\references{
Lele, S.R., Moreno, M. and Bayne, E. 2011. 
Dealing with detection error in site occupancy surveys: 
What can we do with a single survey? 
\emph{Journal of Plant Ecology} (in press)

Solymos, P., Lele, S. R and Bayne, E. 2011.
Conditional likelihood approach for analyzing single visit
abundance survey data in the presence of zero inflation and
detection error. 
\emph{Environmetrics} (in press)
}
\keyword{ package }
\examples{
## FIXME!!! how to open tutorial/vignette?
}
