\name{load_BAM_QPAD}
\alias{load_BAM_QPAD}
\title{
Load BAM QPAD parameter estimates and support functions
}
\description{
Load BAM QPAD parameter estimates and support functions.
}
\usage{
load_BAM_QPAD(version)
}
\arguments{
  \item{version}{
version of the BAM QPAD estimates. List of selection is
provided if missing.
}
}
\value{
The function is called for its side effects: it sources the
BAM QPAD parameter estimates and support functions.
}
\author{
Peter Solymos
}
\keyword{ misc }
